Renamed folder; added Plot.Environment_Grid(); InitializationMethods.Initialize.Fields(); TestPlot
I renamed The Project as there are no more zombies :'(
I added a Plot class with a --> Environment_Grid() Method which can plot
two different types of agents and their health and also shows the sugar
amount . I also added the InitialiyationMethods class in which we should
gather our initialization methods. Last but not least I added file
TetsPlot, which tests the Plot.Environment_Grid(). This one is not jet
working but I will updated it shortly.