plotting does not work
the plotting line in Main does not work so i commented it.
also Viewing-range of Agent is easiest coded in the agent itself, since no mater what information is given to the agent it still needs the exact code part for looking and choosing.
also with the current main (but with "cooler" sugar levels) together with a better human (better then random) and working plotting we have a working model! :D
not sure what "environment" means since gridInfo contains everything needed for both plotting and for the agents movement.